+++ R/scorer-detect.R
@@ -51,7 +51,7 @@
 #'   # create a new Task
 #'   tsk <- Task$new(
 #'     dataset = simple_addition,
-#'     solver = generate(solver_chat = chat_anthropic(model = "claude-3-7-sonnet-latest")),
+#'     solver = generate(solver_chat = chat_anthropic(model = "claude-sonnet-4-5-20250929")),
 #'     scorer = detect_includes()
 #'   )
 #'