+++ inst/regenerate-example-objects.R
@@ -32,7 +32,7 @@ regenerate_example_task <- function() {
     target = c("4", "5")
   )
 
-  ch <- chat_anthropic(model = "claude-3-7-sonnet-latest")
+  ch <- chat_anthropic(model = "claude-sonnet-4-5-20250929")
   tsk <- Task$new(
     dataset = simple_addition,
     solver = generate(ch$set_system_prompt("Respond in the format 'a+b=c'.")),
@@ -48,7 +48,7 @@ regenerate_example_solver <- function() {
 
   solver <- chat_anthropic(
     "Respond in the format 'a+b=c'.",
-    model = "claude-3-7-sonnet-latest"
+    model = "claude-sonnet-4-5-20250929"
   )
   solver$chat("What's 2+2?")