+++ tests/testthat/test-translate-messages.R
@@ -17,7 +17,8 @@ test_that("translate_to_messages works with example turns", {
 
 
 test_that("logs including system prompts are compatible with inspect", {
-  skip_if(identical(Sys.getenv("OPENAI_API_KEY"), ""))
+  vcr::local_cassette("translate-messages-system-prompts")
+  key_get("OPENAI_API_KEY")
   tmp_dir <- withr::local_tempdir()
   withr::local_envvar(list(VITALS_LOG_DIR = tmp_dir))
   withr::local_options(cli.default_handler = function(...) {